Early preview of library amendments to permit core compilation with the new 2.0.0 arduino-esp32 release.

+++ src/libssh_esp32_compat.c
@@ -59,7 +59,7 @@ struct passwd *getpwnam (const char *name)
 { return &p; }
 
 __attribute__((weak))
-int gethostname(char *name, int len)
+int gethostname(char *name, size_t len)
 {
   strncpy(name, LIBSSH_ESP32_COMPAT_HOSTNAME, len);
   return 0;
